requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decouple==0.0.7
distro==1.8.0
exceptiongroup==1.2.0
fastapi==0.104.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.6
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==12.0
